backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9

SQLAlchemy==2.0.35
pydantic==2.9.2
pydantic[email]==2.9.2

passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# audio + whisper
faster-whisper==1.0.3
soundfile==0.12.1
av==12.3.0
numpy==2.1.1

# CORS
